+++ ticket.h
@@ -7,22 +7,22 @@ using namespace std;
 class Ticket
 {
 public:
-	Ticket(const char* date, float price);
+	Ticket(const char* date, double price);
 	virtual ~Ticket();
 
 	//getters
-	float getPrice() const;
+	double getPrice() const;
 	const char* getDate() const;
 
 	//setters
-	void setPrice(float price);
+	void setPrice(double price);
 	void setDate(const char* date);
 
 	//print
 	friend ostream& operator<<(ostream& os, const Ticket& t);
 
 private:
-	float price;
+	double price;
 	char date[11]; //  (dd/mm/yyyy) + '/0' [e.g -  "24/08/2015"]
 };
